refactor: group shared feedback components
