+++ src/shared/components/feedback/Feedback.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Feedback as Template } from "@src/shared/components/Feedback";
+import { Feedback as Template } from "@src/shared/components/feedback/Feedback";
 import { AiOutlineArrowUp } from "react-icons/ai";
 
 const meta = {
