+++ src/shared/components/index.ts
@@ -2,7 +2,7 @@ export * from "@src/shared/components/forms/Button";
 export * from "@src/shared/components/content/Card";
 export * from "@src/shared/components/content/Code";
 export * from "@src/shared/components/content/Description";
-export * from "@src/shared/components/Feedback";
+export * from "@src/shared/components/feedback/Feedback";
 export * from "@src/shared/components/forms/Form";
 export * from "@src/shared/components/forms/FormItem";
 export * from "@src/shared/components/layout/FullScreen";
@@ -14,7 +14,7 @@ export * from "@src/shared/components/content/Logo";
 export * from "@src/shared/components/layout/Main";
 export * from "@src/shared/components/content/Math";
 export * from "@src/shared/components/layout/Outer";
-export * from "@src/shared/components/Overlay";
+export * from "@src/shared/components/feedback/Overlay";
 export * from "@src/shared/components/content/Score";
 export * from "@src/shared/components/content/Section";
 export * from "@src/shared/components/forms/Select";
